+++ source/chip/STM32/peripherals/SPIv2/STM32-SPIv2-spiLowLevelInitializer.cpp
@@ -23,6 +23,17 @@ namespace
 {
 
 /*---------------------------------------------------------------------------------------------------------------------+
+| local objects
++---------------------------------------------------------------------------------------------------------------------*/
+
+/// priority of SPI interrupts
+#if defined(CONFIG_ARCHITECTURE_ARMV7_M_KERNEL_BASEPRI)
+constexpr uint8_t interruptPriority {CONFIG_ARCHITECTURE_ARMV7_M_KERNEL_BASEPRI};
+#else	// !defined(CONFIG_ARCHITECTURE_ARMV7_M_KERNEL_BASEPRI)
+constexpr uint8_t interruptPriority {};
+#endif	// !defined(CONFIG_ARCHITECTURE_ARMV7_M_KERNEL_BASEPRI)
+
+/*---------------------------------------------------------------------------------------------------------------------+
 | local functions
 +---------------------------------------------------------------------------------------------------------------------*/
 
@@ -35,27 +46,27 @@ namespace
 void spiLowLevelInitializer()
 {
 #ifdef CONFIG_CHIP_STM32_SPIV2_SPI1_ENABLE
-	NVIC_SetPriority(SPI1_IRQn, {});
+	NVIC_SetPriority(SPI1_IRQn, interruptPriority);
 	NVIC_EnableIRQ(SPI1_IRQn);
 #endif	// def CONFIG_CHIP_STM32_SPIV2_SPI1_ENABLE
 #ifdef CONFIG_CHIP_STM32_SPIV2_SPI2_ENABLE
-	NVIC_SetPriority(SPI2_IRQn, {});
+	NVIC_SetPriority(SPI2_IRQn, interruptPriority);
 	NVIC_EnableIRQ(SPI2_IRQn);
 #endif	// def CONFIG_CHIP_STM32_SPIV2_SPI2_ENABLE
 #ifdef CONFIG_CHIP_STM32_SPIV2_SPI3_ENABLE
-	NVIC_SetPriority(SPI3_IRQn, {});
+	NVIC_SetPriority(SPI3_IRQn, interruptPriority);
 	NVIC_EnableIRQ(SPI3_IRQn);
 #endif	// def CONFIG_CHIP_STM32_SPIV2_SPI3_ENABLE
 #ifdef CONFIG_CHIP_STM32_SPIV2_SPI4_ENABLE
-	NVIC_SetPriority(SPI4_IRQn, {});
+	NVIC_SetPriority(SPI4_IRQn, interruptPriority);
 	NVIC_EnableIRQ(SPI4_IRQn);
 #endif	// def CONFIG_CHIP_STM32_SPIV2_SPI4_ENABLE
 #ifdef CONFIG_CHIP_STM32_SPIV2_SPI5_ENABLE
-	NVIC_SetPriority(SPI5_IRQn, {});
+	NVIC_SetPriority(SPI5_IRQn, interruptPriority);
 	NVIC_EnableIRQ(SPI5_IRQn);
 #endif	// def CONFIG_CHIP_STM32_SPIV2_SPI5_ENABLE
 #ifdef CONFIG_CHIP_STM32_SPIV2_SPI6_ENABLE
-	NVIC_SetPriority(SPI6_IRQn, {});
+	NVIC_SetPriority(SPI6_IRQn, interruptPriority);
 	NVIC_EnableIRQ(SPI6_IRQn);
 #endif	// def CONFIG_CHIP_STM32_SPIV2_SPI6_ENABLE
 }